requirements.txt
numpy
matplotlib
Pillow
picamera
pyserial
rpi.gpio
ffmpeg-python